+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,10 @@
         android:supportsRtl="true"
         tools:targetApi="31"
         android:theme="@style/PrimaryThemes"
+        android:usesCleartextTraffic="true"
         >
         <activity
-            android:name=".MainActivity"
+            android:name="com.emedia.dprri.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/PrimaryThemes"
